refactor: create userstats api for interaction with all userstats
